# api/openapi.yaml
openapi: 3.1.0
info:
  title: BugBarn API
  description: |
    Self-hosted error tracking, structured logging, and web analytics.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    - **Session cookie** — Log in via `POST /api/v1/login` to receive a `bugbarn_session` cookie.
      State-changing requests (POST, PUT, PATCH, DELETE) must include the `X-BugBarn-CSRF` header
      with the value from the `bugbarn_csrf` cookie.
    - **API key** — Pass a key in the `X-BugBarn-API-Key` header. Keys have scopes:
      `full` (all access), `read` (GET only, no settings/apikeys), or `ingest` (POST to ingest endpoints only).

    Ingest endpoints (`/api/v1/events`, `/api/v1/logs`, `/api/v1/analytics/collect`) use
    wildcard CORS and accept ingest-scoped API keys so browser SDKs can POST from any origin.
  version: 1.0.0
  contact:
    name: BugBarn
    url: https://bugbarn.dev
servers:
  - url: https://bugbarn.wiebe.xyz
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /api/v1/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Authenticate and create a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Session and CSRF cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  authEnabled:
                    type: boolean
                  username:
                    type: string
        "401":
          description: Invalid credentials
        "429":
          description: Too many login attempts
          headers:
            Retry-After:
              schema:
                type: string
...
  /api/v1/oidc/backchannel-logout:
    post:
      operationId: oidcBackchannelLogout
      tags: [Auth]
      summary: OIDC Back-Channel Logout 1.0 endpoint
      description: |
        Called server-to-server by the identity provider when a user's IdP
        session ends (logout, admin revocation, suspension). The signed
        `logout_token` is verified against the issuer's JWKS; matching local
        sessions (by `sid`, else by `sub`) are deleted immediately.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [logout_token]
              properties:
                logout_token:
                  type: string
                  description: Signed JWT logout token minted by the IdP
      responses:
        "200":
          description: Matching sessions terminated (also returned when none matched)
        "400":
          description: Invalid logout token
        "429":
          description: Rate limited
  /api/v1/me:
    get:
      operationId: me
      tags: [Auth]
      summary: Get current user info
      security: []
      responses:
        "200":
          description: Authenticated user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  authEnabled:
                    type: boolean
                  username:
                    type: string
                  authMethod:
                    type: string
                    enum: [oidc, local]
        "401":
          description: Not authenticated

  # ──────────────────────────────────────────────
  # Ingest (wildcard CORS)
  # ──────────────────────────────────────────────
  /api/v1/events:
    post:
      operationId: ingestEvent
      tags: [Ingest]
      summary: Ingest an error event
      description: |
        Accepts error events from SDKs. Uses wildcard CORS so browser SDKs can POST
        from any origin. Accepts ingest-scoped or full-scoped API keys.

        The response status draws a hard line between accepted and dropped events:

        - **2xx** — accepted: the event was validated and durably queued, and
          will be processed. The body is `IngestAccepted`.
        - **4xx (400, 401, 413)** — dropped permanently: the request is bad and
          will never succeed as-is. Do **not** retry.
        - **429 / 503** — dropped transiently: backpressure or a temporary
          outage. Retry after the `Retry-After` delay.

        Dropped responses carry an `IngestDropped` body whose `retryable` flag
        restates whether retrying the identical request could succeed. Events are
        validated synchronously, so a `202` means the event was genuinely
        accepted rather than queued and later discarded.
      security:
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/XBugBarnProject"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestEvent"
      responses:
        "202":
          description: Event accepted — durably queued and will be processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestAccepted"
        "400":
          description: Dropped — malformed event payload (permanent, do not retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestDropped"
        "401":
          description: Dropped — missing or invalid API key (permanent, do not retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestDropped"
        "413":
          description: Dropped — payload exceeds the size limit (permanent, do not retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestDropped"
        "429":
          description: Dropped — ingest buffer full (transient, retry after Retry-After)
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestDropped"
        "503":
          description: Dropped — ingest pipeline unavailable (transient, retry after Retry-After)
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestDropped"
  /api/v1/logs:
    post:
      operationId: ingestLogs
      tags: [Ingest]
      summary: Ingest structured logs
      description: |
        Accepts structured log entries. Supports JSON array or NDJSON format.
        Uses wildcard CORS. Requires a project context via API key or X-BugBarn-Project header.
      security:
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/XBugBarnProject"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logs:
                  type: array
                  items:
                    $ref: "#/components/schemas/LogObject"
          application/x-ndjson:
            schema:
              type: string
              description: Newline-delimited JSON log objects
      responses:
        "204":
          description: Logs accepted
        "400":
          description: Invalid payload or missing project
    get:
      operationId: listLogs
      tags: [Logs]
      summary: List structured logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
        - name: level
          in: query
          schema:
            type: string
            enum: [trace, debug, info, warn, error, fatal]
          description: Minimum log level
        - name: q
          in: query
          schema:
            type: string
          description: Full-text search query
        - name: before
          in: query
          schema:
            type: integer
            format: int64
          description: Cursor for pagination (log entry ID)
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/LogEntry"
                  next_cursor:
                    type: integer
                    format: int64
  /api/v1/analytics/collect:
    post:
      operationId: collectPageView
      tags: [Ingest]
      summary: Collect a page view for web analytics
      description: |
        Accepts page view data from the analytics snippet. Uses wildcard CORS
        for sendBeacon compatibility. No API key required.
      security: []
      parameters:
        - name: project
          in: query
          schema:
            type: string
          description: Project slug
        - $ref: "#/components/parameters/XBugBarnProject"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageViewPayload"
      responses:
        "202":
          description: Page view accepted
        "400":
          description: Invalid payload or missing pathname

  # ──────────────────────────────────────────────
  # Analytics JS snippet
  # ──────────────────────────────────────────────
  /analytics.js:
    get:
      operationId: getAnalyticsSnippet
      tags: [Analytics]
      summary: Serve the analytics JavaScript snippet
      security: []
      parameters:
        - name: project
          in: query
          schema:
            type: string
          description: Project slug to embed in the snippet
      responses:
        "200":
          description: JavaScript snippet
          content:
            application/javascript:
              schema:
                type: string

  # ──────────────────────────────────────────────
  # Setup
  # ──────────────────────────────────────────────
...
  /api/v1/issues:
    get:
      operationId: listIssues
      tags: [Issues]
      summary: List issues with filtering and pagination
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [last_seen, first_seen, event_count]
          description: Sort order
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, muted, all]
          description: Filter by status
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: project_slug
          in: query
          schema:
            type: string
          description: Filter by project slug
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  issues:
                    type: array
                    items:
                      $ref: "#/components/schemas/Issue"
                  hasMore:
                    type: boolean
  /api/v1/issues/sparklines:
    get:
      operationId: issueSparklines
      tags: [Issues]
      summary: Get 24-hour event count sparklines for issues
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated issue display IDs (e.g. "BW-1,BW-2")
      responses:
        "200":
          description: Sparkline data keyed by display ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  sparklines:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: integer
                      minItems: 24
                      maxItems: 24
...
  /api/v1/events/{eventId}:
    get:
      operationId: getEvent
      tags: [Events]
      summary: Get a single event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
  /api/v1/events/stream:
    get:
      operationId: streamEvents
      tags: [Events]
      summary: SSE stream of live error events
      description: |
        Server-Sent Events stream that emits new error events as they are processed.
        Polls every 3 seconds and sends keepalive comments every 15 seconds.
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only emit events after this timestamp (RFC 3339)
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
  /api/v1/projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                slug:
                  type: string
                  description: Auto-generated from name if omitted
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
...
  /api/v1/projects/{slug}:
    put:
      operationId: renameProject
      tags: [Projects]
      summary: Rename a project
      parameters:
        - $ref: "#/components/parameters/projectSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
      responses:
        "200":
          description: Project renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a project and all its data
      parameters:
        - $ref: "#/components/parameters/projectSlug"
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/v1/projects/{slug}/approve:
    post:
      operationId: approveProject
      tags: [Projects]
      summary: Approve a pending project
      parameters:
        - $ref: "#/components/parameters/projectSlug"
      responses:
        "200":
          description: Project approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/v1/projects/{slug}/merge:
    post:
      operationId: mergeProject
      tags: [Projects]
      summary: Merge a project into another
      parameters:
        - $ref: "#/components/parameters/projectSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target:
                  type: string
                  description: Target project slug to merge into
      responses:
        "200":
          description: Project merged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ──────────────────────────────────────────────
  # Groups
  # ──────────────────────────────────────────────
  /api/v1/groups:
    get:
      operationId: listGroups
      tags: [Groups]
      summary: List project groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectGroup"
    post:
      operationId: createGroup
      tags: [Groups]
      summary: Create a project group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                slug:
                  type: string
                  description: Auto-generated from name if omitted
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    $ref: "#/components/schemas/ProjectGroup"
...
  /api/v1/releases:
    get:
      operationId: listReleases
      tags: [Releases]
      summary: List releases
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                type: object
                properties:
                  releases:
                    type: array
                    items:
                      $ref: "#/components/schemas/Release"
    post:
      operationId: createRelease
      tags: [Releases]
      summary: Create a release marker
      description: >
        Accepts a release marker and enqueues it for asynchronous creation by the
        background worker, returning 202 immediately. This keeps the request off
        the single SQLite writer connection. The marker is typically persisted
        within ~1s and becomes visible via GET /api/v1/releases shortly after.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseInput"
      responses:
        "202":
          description: Release marker accepted for asynchronous creation
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  ingestId:
                    type: string
...
  /api/v1/alerts:
    get:
      operationId: listAlerts
      tags: [Alerts]
      summary: List alert rules
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [Alerts]
      summary: Create an alert rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "200":
          description: Alert created
          content:
            application/json:
              schema:
                type: object
                properties:
                  alert:
                    $ref: "#/components/schemas/Alert"
...
  /api/v1/logs/stream:
    get:
      operationId: streamLogs
      tags: [Logs]
      summary: SSE stream of live log entries
      description: |
        Server-Sent Events stream that emits new log entries as they arrive.
        Sends keepalive comments every 15 seconds.
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "503":
          description: Log streaming unavailable

  # ──────────────────────────────────────────────
  # Settings
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bugbarn_session
      description: |
        Session cookie set by POST /api/v1/login. State-changing requests
        must also include the X-BugBarn-CSRF header with the value from
        the bugbarn_csrf cookie.
    apiKey:
      type: apiKey
      in: header
      name: X-BugBarn-API-Key
      description: |
        API key header. Scopes: `full` (all access), `read` (GET only, excludes
        settings and apikeys), `ingest` (POST to ingest endpoints only).
...